Move CommitNode/CommitNodeIndex into separate object/commitgraph package
