+++ plumbing/object/commitgraph/commitnode_test.go
@@ -1,4 +1,4 @@
-package object
+package commitgraph
 
 import (
 	"path"
