+++ plumbing/object/commitgraph/commitnode_walker_ctime.go
@@ -1,4 +1,4 @@
-package object
+package commitgraph
 
 import (
 	"io"
